change GanttChar

+++ PL/PO/TaskGantt.cs
@@ -6,7 +6,7 @@ public class TaskGantt
     public string taskName { get; init; }
     public BO.Status taskStatus { get; init; }
     public int duration { get; set; }    //width of rectangle task
-    public int timeFromStart { get; set; }  //width of empty rectangle from start to task
-    public int timeToEnd { get; set; }    //width of empty rectangle from task to end
+    public double timeFromStart { get; set; }  //width of empty rectangle from start to task
+    public double timeToEnd { get; set; }    //width of empty rectangle from task to end
 }
